backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.12

# Pydantic
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Vector Database
pymilvus>=2.4.0

# Embedding & ML
sentence-transformers>=2.3.0
torch>=2.0.0
transformers>=4.36.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# HTTP Client
httpx>=0.26.0
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Database
pymongo>=4.6.0
bson>=0.5.10

# Google Drive API
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Async Support
aiofiles>=23.2.0

# LangChain Full Integration
langchain>=0.1.0
langchain-core>=0.1.23
langchain-community>=0.0.20
langchain-openai>=0.1.0
langchain-text-splitters>=0.0.1
langchain-milvus>=0.1.0

# Authentication
PyJWT>=2.8.0

# Data Processing
numpy>=1.24.0
tqdm>=4.65.0

# Progress Bar
tqdm>=4.65.0